# protocol/openapi.yaml
openapi: 3.0.3
info:
  title: Remote File Explorer — Host Agent API
  version: 1.17.0
  description: >
    Contract between the Flutter mobile app and the Go host agent. JSON for control and
    metadata; binary streams for file content. All routes except /pair, /register,
    /login, /auth/challenge, and /health require a bearer device token. Clients may
    send an optional `X-RFE-Client-Version` header (e.g.
    "1.10.0+18") on every authenticated request; the agent records this plus the
    caller's network address against the device row. Admin devices see those fields
    for all devices via GET /devices (see Device.lastAddress / Device.lastVersion);
    ordinary paired devices receive only their own minimal device record.

    A device's token is "admin" (Device.viaLogin=true) if it was obtained via /login or
    /register (proof of the single account's password) rather than /pair (a one-time
    code meant for an ordinary device like the phone app). Admin devices may manage
    OTHER devices: POST /pairing/generate, and PATCH/DELETE /devices/{id} targeting a
    device other than themselves. An ordinary (non-admin) device may still manage only
    ITSELF via DELETE /devices/{id}.

    Pair, register, and login all require device-identity proof-of-possession: fetch a
    nonce from POST /auth/challenge, sign it with the device's permanent Ed25519 key,
    and send the (devicePublicKey, nonce, signature) triple. The agent pins the public
...
servers:
...
    variables:
...
security:
  - deviceToken: []
paths:
...
  /status:
    get:
      summary: Agent self-report (uptime, version, platform, disk space)
      security: [{ deviceToken: [] }]
      responses:
        '200':
          description: Agent status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentStatus' }
        '401': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
...
  /transfers/list:
    get:
      summary: List recent upload transfer sessions (web companion Transfers page)
      description: >
        Read-only listing of the chunked-upload sessions the agent tracks.
        The transfers array is capped at the most recent 200 rows (the table
        accumulates every session ever opened); counts/total aggregate the
        whole table so the summary cards stay accurate. progress is derived
        from received-chunk count; "open" means started-but-never-finalized.
        Distinct from GET /transfers/{id} (single-session status) and POST
        /transfers (open a session).

        Scoped to the caller: an admin device (authenticated via /login or
        /register) sees every session plus the whole-table counts and the
        devices/users filter lists. A code-paired device sees only the
        sessions it opened, its device/user query params are ignored rather
        than honoured, counts/activeNow are derived from its own rows (and so
        capped at the same 200), and devices/users come back empty.
      security: [{ deviceToken: [] }]
      parameters:
        - name: device
          in: query
          required: false
          schema: { type: string }
          description: >
            Restrict results to transfers opened by this device ID. Omit or
            pass empty for all devices. Values come from the devices list in
            this endpoint's own response. Admin only — ignored for a
            code-paired caller, which is always scoped to itself.
        - name: user
          in: query
          required: false
          schema: { type: string }
          description: >
            Restrict results to transfers opened by devices authenticated as
            this login account (devices are stamped with a username at
            /login and /register; pair-flow devices are unattributed). Values
            come from the users list in this endpoint's own response.
            Combines with the device filter (AND).
      responses:
        '200':
          description: Recent transfer sessions plus whole-table status counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer, description: Total transfer rows in the table }
                  counts:
                    type: object
                    additionalProperties: { type: integer }
                    description: Row count per status (e.g. open, completed, failed)
                  activeNow:
                    type: integer
                    description: Open transfers that received a chunk in the last 30s
                  devices:
                    type: array
                    description: Distinct devices with at least one transfer, for the filter-chip row and the table's Device/User columns
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        label: { type: string }
                        username: { type: string, description: "\"\" if the device paired by code rather than login" }
                  users:
                    type: array
                    description: Distinct login accounts with at least one transfer, for the user filter-chip row
                    items: { type: string }
                  transfers:
                    type: array
                    items: { $ref: '#/components/schemas/TransferSummary' }
        '401': { $ref: '#/components/responses/Error' }
        '403':
          description: >
            Caller has no device identity at all. code=FORBIDDEN.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /users:
    get:
      summary: List login accounts (web companion Users page)
      description: >
        Usernames and creation times only — never password hashes. RFE has no
        per-user roles or session tracking, so those Figma-style columns are
        intentionally absent.

        Admin only: a code-paired device (the phone app, a guest) gets 403 and
        cannot enumerate the host's accounts.
      security: [{ deviceToken: [] }]
      responses:
        '200':
          description: Login accounts
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/UserSummary' } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /users/{username}:
    delete:
      summary: Permanently remove a login account (web companion Users page)
      description: >
        Admin only — a code-paired device gets 403. Every login account is a
        full admin today (see the /users GET note), so there's no
        self-vs-other distinction like device management has: any admin may
        remove any account. The one exception: the LAST remaining account
        can't be deleted (code=LAST_USER), since that would brick password
        login entirely (no way to obtain a device token except an existing
        pairing code).
      security: [{ deviceToken: [] }]
      parameters:
        - { name: username, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Account removed }
        '400':
          description: >
            username is the only remaining account. code=LAST_USER.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404':
          description: No account with that username. code=NOT_FOUND.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /logs:
    get:
      summary: Tail the agent's own log output (web companion Logs page)
      description: >
        Last ~200 lines of the agent's journald output (Linux/systemd). Empty
        array on non-systemd hosts. RFE logs are unstructured, so there is no
        severity level field.

        Admin only: the journal spans every device, so a code-paired caller
        (which may itself be jailed or read-only) gets 403.
      security: [{ deviceToken: [] }]
      responses:
        '200':
          description: Recent log lines
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/LogLine' } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /audit:
    get:
      summary: Read the account/device/share audit trail
      description: >
        Security-relevant events only — pairing, registration, login (including
        failed attempts), device revoke/remove/limit changes, share-link
        mint/revoke, and agent restart. File operations are deliberately not
        recorded: they arrive at transfer volume and would bury everything
        else. The log keeps the most recent 5000 entries.

        Admin only: it spans every device, so a code-paired caller gets 403.
      security: [{ deviceToken: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100, maximum: 500 }
          description: Entries per page. Out-of-range values fall back to 100.
        - in: query
          name: before
          schema: { type: integer, format: int64 }
          description: >
            Page backwards by passing the `id` of the last entry from the
            previous response. Omit for the newest page.
      responses:
        '200':
          description: Audit entries, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEntry' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /agent/restart:
    post:
      summary: Restart the agent's own service (Linux/systemd only)
      description: >
        Admin only — a code-paired device gets 403, since a restart cuts every
        other device's transfers. Restart-only, no remote stop, so there is
        nothing to get permanently stuck in. Shells out to
        `systemctl --user restart rfe-agent.service`; the response is sent
        before the command runs. Returns 501 on non-Linux agents.
      security: [{ deviceToken: [] }]
      responses:
        '202':
          description: Restart triggered
          content:
            application/json:
              schema: { type: object, properties: { ok: { type: boolean } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '501':
          description: Not supported on this platform. code=NOT_IMPLEMENTED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /auth/challenge:
    post:
      summary: Mint a single-use nonce for device-identity proof-of-possession (unauthenticated)
      description: >
        Sign the returned nonce with the device's Ed25519 private key and pass
        it (with the device's public key) to /pair, /register, or /login as
        proof the caller controls that key. Valid for 2 minutes, single-use.
      security: []
      responses:
        "200":
          description: Nonce minted
          content:
            application/json:
              schema:
                type: object
                properties: { nonce: { type: string } }
        "429":
          description: Too many challenge requests (rate-limited at 30/min). code=RATE_LIMITED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /pair:
    post:
      summary: Complete device pairing (unauthenticated; consumes one-time pairing code)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PairRequest" }
      responses:
        "200":
          description: Device paired; token issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairResponse" }
        "400":
          description: >
            Malformed body, or missing devicePublicKey/nonce/signature.
            code=BAD_REQUEST or code=DEVICE_KEY_REQUIRED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "401":
          description: >
            Invalid pairing code (code=INVALID_CODE), nonce missing/expired/reused
            (code=INVALID_NONCE), or signature doesn't verify (code=INVALID_SIGNATURE).
            A key change for an already-known device id is accepted here (re-pinned),
            not rejected — see the device-identity note above.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: >
            Too many pairing attempts (rate-limited at 10/min per agent, since the
            pairing code is short and brute-forceable). code=RATE_LIMITED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /register:
    post:
      summary: >
        Create the account and pair this device in one step (unauthenticated;
        consumes one-time pairing code). An alternative to `rfe-agent adduser`
        for first-time setup entirely from the phone app or web companion.
        Requires the same one-time pairing code as POST /pair — this is what
        stops a stranger on the network from creating an account before the
        owner does. Enforces one account per agent: fails once any account
        already exists (code=ACCOUNT_ALREADY_EXISTS), regardless of pairing
        code validity — log in instead. Response shape is identical to
        PairResponse.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "200":
          description: Account created, device paired; token issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairResponse" }
        "400":
          description: >
            Malformed body, missing username, password under 8 characters, or
            missing devicePublicKey/nonce/signature. code=BAD_REQUEST or
            code=DEVICE_KEY_REQUIRED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "401":
          description: >
            Invalid pairing code (code=INVALID_CODE), or device-identity proof
            failed (code=INVALID_NONCE / INVALID_SIGNATURE — see POST /pair).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: >
            This agent already has an account (code=ACCOUNT_ALREADY_EXISTS —
            one account per agent; log in instead), or that exact username
            was taken by a concurrent request (code=USERNAME_TAKEN).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: >
            Too many registration attempts (rate-limited at 10/min per agent).
            code=RATE_LIMITED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /login:
    post:
      summary: >
        Log in with the account created via POST /register or `rfe-agent
        adduser` (unauthenticated). An additional way to obtain a device
        token alongside POST /pair — not a replacement for it. Response
        shape is identical to PairResponse.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in; device token issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairResponse" }
        "400":
          description: >
            Malformed body, missing username/password, or missing
            devicePublicKey/nonce/signature. code=BAD_REQUEST or
            code=DEVICE_KEY_REQUIRED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "401":
          description: >
            Invalid username or password (code=INVALID_CREDENTIALS — the same
            error is returned for a nonexistent username as for a wrong
            password, so a caller can't enumerate valid usernames), or
            device-identity proof failed (code=INVALID_NONCE / INVALID_SIGNATURE
            / DEVICE_KEY_MISMATCH — see POST /pair).
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: >
            Too many login attempts (rate-limited at 10/min per agent).
            code=RATE_LIMITED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /wol:
    post:
      summary: Send a Wake-on-LAN magic packet to a MAC address on the agent's LAN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mac]
              properties:
                mac: { type: string, description: "Colon-separated MAC address (e.g. aa:bb:cc:dd:ee:ff)" }
      responses:
        "200":
          description: Magic packet sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [sent] }
        "400": { $ref: "#/components/responses/Error" }
  /share/mint:
    post:
      summary: Mint a one-time share link for a single file
      description: >
        Generates a short-lived, single-use, unauthenticated download URL for
        one file. Returns 403 if the agent's `allowSharing` setting is
        disabled (default). Only regular files up to 500 MiB are accepted.
        See docs/r1-share-link-threat-model.md.
      security: [{ deviceToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path: { type: string, description: "Absolute path to the file to share." }
                expiresInSeconds:
                  type: integer
                  format: int64
                  description: "Defaults to 900 (15min) if absent/zero. Maximum 86400 (24h)."
      responses:
        "200":
          description: Share link minted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShareLink" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: Share links are disabled on this agent. code=FORBIDDEN.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "404": { $ref: "#/components/responses/Error" }
        "413":
          description: The file exceeds the 500 MiB share limit. code=FILE_TOO_LARGE.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  /share:
    get:
      summary: List active (unexpired) share links
      description: >
        Scoped to the caller: an admin device (authenticated via /login or
        /register) sees every active link; a code-paired device sees only the
        links it minted. Links minted before the agent recorded owners appear
        for admins only.
      security: [{ deviceToken: [] }]
      responses:
        "200":
          description: Active share links
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ShareLinkSummary" }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: >
            Caller has no device identity at all. code=FORBIDDEN.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  /fs/folder:
    post:
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequest" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Entry" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /fs/file:
    post:
      summary: Create an empty file
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequest" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Entry" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
...
  schemas:
...
    AgentStatus:
      type: object
      properties:
        version: { type: string }
        uptimeSeconds: { type: integer, format: int64 }
        platform: { type: string }
        freeBytes: { type: integer, format: int64 }
        totalBytes: { type: integer, format: int64 }
...